refactor: extract TranscriptionChunker.swift (#648 slice E FINAL)
Fifth and final slice of #648 (TranscriptionClient.swift decomposition).
Extracts TranscriptionAudioChunk (Sendable struct) + DefaultTranscriptionChunker
(TranscriptionChunking conformance) into a new TranscriptionChunker.swift.

Byte-identical move; SHA-256 4ca276eb8a4c7f506293e916d806a9c54f78789553b9e86b426193a40fdf34ac
verified against origin/main:TranscriptionClient.swift lines 483-614.

Imports: @preconcurrency AVFoundation (for AVURLAsset, AVAssetExportSession)
+ Foundation.

Reduces TranscriptionClient.swift from 614 to 482 lines. After all 5 slices of
#648, TranscriptionClient.swift went from 753 lines to ~482 lines (~36%
reduction), containing only the primary TranscriptionClient actor.

Tests: 667 (unchanged).
